remove the aggregation folder and move aggregation into transaction
