--- arm/src/aggregation/mod.rs
@@ -1,3 +1,0 @@
-//! Aggregation module for ARM proofs.
-pub mod batch;
-pub mod constants;
