+++ arm/src/lib.rs
@@ -5,8 +5,6 @@
 #[cfg(feature = "transaction")]
 pub mod action;
 pub mod action_tree;
-#[cfg(feature = "aggregation")]
-pub mod aggregation;
 #[cfg(any(feature = "compliance_circuit", feature = "aggregation_circuit"))]
 pub mod compliance;
 #[cfg(feature = "transaction")]
